Relax related records matching for comments in included resources specs
Specs deeper down the complex spec of included resources change what
records CommentPolicy::Scope returns. This in turn affects our ability
to expect the exact objects passed to the authorizer.

We're fine with relaxing the constraint in these specs as the exact
models are tested in the create_resource callback test